Reorganize index remaining dumps job
